% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{as_eth_date}
\alias{as_eth_date}
\alias{as_eth_date.numeric}
\alias{as_eth_date.character}
\alias{as_eth_date.Date}
\title{Date Conversion Functions to Ethiopian Date}
\usage{
as_eth_date(x, ...)

\method{as_eth_date}{numeric}(x, ...)

\method{as_eth_date}{character}(x, ..., sep = "-", orders = "ymd")

\method{as_eth_date}{Date}(x, ...)
}
\arguments{
\item{x}{an integer, character, or date objects to be parsed as an Ethiopian date}

\item{...}{Arguments that passes to and from methods. Must be empty.}

\item{sep}{if `x` is a character, an separator of year, month and day.}

\item{orders}{if `x` is a character, the order of year (y), month (m) and day (d)
in the character. Default is "ymd".}
}
\value{
A vector of "ethDate" class.
}
\description{
These functions can parse an Ethiopian date from integers, and characters.
It can seamlessly convert Ethiopian date to and from Gregorian calendar.
At the moment, there's no function that supports timezone and hours.
}
\examples{
as_eth_date("2000-01-01")
}
\author{
Gutama Girja Urago
}
