% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cpieth}
\alias{cpieth}
\title{Monthly FAO Food Price Index 2001-2023}
\format{
\subsection{\code{cpieth}}{

A data frame with 270 rows and 2 columns:
\describe{
\item{date}{Date (GC)}
\item{cpi}{Consumer price index}
}
}
}
\source{
\url{https://data360.worldbank.org/en/indicator/FAO_CP_23014?country=ETH}
}
\usage{
cpieth
}
\description{
A subset of data from the FAO Food Price Index for Ethiopia.
}
\keyword{datasets}
