% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{eth_make_date}
\alias{eth_make_date}
\title{Make Ethiopian Date}
\usage{
eth_make_date(year, month, day)
}
\arguments{
\item{year}{an integer vector of Ethiopian year.}

\item{month}{an integer vector of Ethiopian month.}

\item{day}{an integer vector of Ethiopian day.}
}
\value{
a vector of 'ethdate' objects.
}
\description{
Make Ethiopian date from year, month and day components.
}
\details{
This function makes an Ethiopian date object from three integer vectors of an equal length.
It validates the date and returns \code{NA} for invalid dates. It accounts for leap years.
}
\examples{
eth_make_date(2017, 01, 15)
}
\seealso{
\code{\link[=eth_date]{eth_date()}} \code{\link[=eth_parse_date]{eth_parse_date()}}
}
\author{
Gutama Girja Urago
}
