% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{eth_show}
\alias{eth_show}
\alias{eth_today}
\alias{eth_now}
\title{See Month or Day Names}
\usage{
eth_show(x = c("\%B", "\%b", "\%A", "\%a"), lang = c("lat", "amh", "en"))

eth_today(...)

eth_now(...)
}
\arguments{
\item{x}{what you want to see.}

\item{lang}{language of the text.}

\item{...}{arguments that passes to \code{\link[=format]{format()}}}
}
\value{
Except for \code{eth_date()}, which returns an \code{ethdate} object, other functions return a character vector.
}
\description{
Small functions that displays texts.
}
\details{
\code{eth_show()} displays the underlying month and weekday names that is used by \code{\link[=eth_parse_date]{eth_parse_date()}}.
}
\examples{
eth_show()
eth_show("\%A", "amh")
eth_today()
eth_now()

}
\author{
Gutama Girja Urago
}
