% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{eth_year}
\alias{eth_year}
\alias{eth_month}
\alias{eth_day}
\alias{eth_weekday}
\alias{eth_monthname}
\title{Extract Components of Dates}
\usage{
eth_year(x)

eth_month(x)

eth_day(x)

eth_weekday(x)

eth_monthname(x)
}
\arguments{
\item{x}{a vector of class Ethiopian dates}
}
\value{
An integer vector
}
\description{
Extract Components of Dates
}
\examples{
if (FALSE) {
x <- as_eth_date("2017-01-01")
year <- eth_year(x)
month <- eth_month(x)
day <- eth_day(x)
weekday <- eth_weekday(x)
monthname <- eth_monthname(x)
}
}
\author{
Gutama Girja Urago
}
