% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{to_gregorian}
\alias{to_gregorian}
\alias{to_ethiopian}
\alias{as_gre_date}
\alias{as.Date.ethDate}
\alias{as_numeric}
\title{Date Conversion Functions to and from Ethiopian Date}
\usage{
to_gregorian(x, ...)

to_ethiopian(x)

as_gre_date(x, ...)

\method{as.Date}{ethDate}(x, ...)

as_numeric(x, ...)
}
\arguments{
\item{x}{a Date or ethDate object}

\item{...}{Arguments that passes to and from methods. Must be empty.}
}
\value{
Except `as_numeric`, return an object with "Date" class if converted to Gregorian,
else an object with "ethDate" class. `as_numeric` returns number of days since 1970-01-01,
which can easily be converted to Gregorian date using `as.Date`.
}
\description{
Functions to convert an Ethiopian date to and from Gregorian calendar.
At the moment, there's no function that supports timezone and hours.
}
\examples{
date <- as.Date("2025-01-01")
eth_date <- to_ethiopian(date)
}
\author{
Gutama Girja Urago
}
