% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{eth_date}
\alias{eth_date}
\title{Create Ethiopian Dates}
\usage{
eth_date(x)
}
\arguments{
\item{x}{a numeric, Date, POSIXct or POSIXt vector used to construct an Ethiopian date.}
}
\value{
Returns a vector with an 'ethDate' class that can further be used as an Ethiopian date.
}
\description{
Convert or create an Ethiopian date.
}
\details{
\code{eth_date()} internally stores number of days since 1962-04-23 EC (equivalent to 1970-01-01 GC).
Days before 1962-04-23 EC are represented as negative integers.
This makes it easy to convert from and to base \code{Date} objects. This function does not
support time and timezone.
}
\examples{
eth_date(as.Date(0))
}
\seealso{
\code{\link[=eth_make_date]{eth_make_date()}}
}
\author{
Gutama Girja Urago
}
