% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{eth_make_date}
\alias{eth_make_date}
\title{Make Ethiopian Date}
\usage{
eth_make_date(year, month, day)
}
\arguments{
\item{year}{an integer vector of Ethiopian year.}

\item{month}{an integer vector of Ethiopian month.}

\item{day}{an integer vector of Ethiopian day.}
}
\value{
Returns a vector with an 'ethDate' class that can further be used as an Ethiopian date.
}
\description{
Make Ethiopian date object from year, month and day.
}
\details{
This function constructs an Ethiopian date object from three vectors of an equal length.
It validates the date and returns \code{NA} for invalid dates. It accounts for leap years.
The returned object can further be used.
If you want to convert it to Gregorian calendar, use \code{as.Date()}.
}
\examples{
eth_make_date(2017, 01, 15)
}
\seealso{
\code{\link[=eth_date]{eth_date()}}
}
\author{
Gutama Girja Urago
}
