% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{eth_parse_date}
\alias{eth_parse_date}
\title{Parse Ethiopian Date}
\usage{
eth_parse_date(x, format = "\%Y-\%m-\%d", lang = c("amh", "lat", "en"))
}
\arguments{
\item{x}{a character vector.}

\item{format}{a format in in which x is composed.}

\item{lang}{a language in which month names are written, if included in x. Use "amh" for month names written in Amharic alphabets,
"lat" for Amharic month names written in Latin alphabets, and "en" for English month names.}
}
\value{
a vector of 'ethdate' objects.
}
\description{
Parse Ethiopian date from character vector that has non-digit separator.
}
\details{
x must include non-digit separator.
}
\examples{
eth_parse_date("2017-01-01")
}
\seealso{
\code{\link[=eth_date]{eth_date()}} \code{\link[=eth_make_date]{eth_make_date()}}
}
\author{
Gutama Girja Urago
}
