% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R, R/utils.R
\name{format.ethdate}
\alias{format.ethdate}
\alias{is_eth_date}
\alias{is_eth_leap}
\alias{as.Date.ethdate}
\alias{as.double.ethdate}
\alias{as.integer.ethdate}
\alias{as.character.ethdate}
\title{#' @export
print.ethdifftime <- function(x, max = NULL, ...) {
x <- as.difftime(vctrs::vec_data(x), units = "days")
print(x, max = max, ...)
}}
\usage{
\method{format}{ethdate}(x, format = "\%Y-\%m-\%d", lang = c("lat", "amh", "en"), ...)

is_eth_date(x)

is_eth_leap(x)

\method{as.Date}{ethdate}(x, ...)

\method{as.double}{ethdate}(x, ...)

\method{as.integer}{ethdate}(x, ...)

\method{as.character}{ethdate}(x, ...)
}
\arguments{
\item{x}{an ethdate or numeric vector.}

\item{format}{a format for character date.}

\item{lang}{a language.}

\item{...}{further arguments to be passed to specific methods.}
}
\value{
\code{is_eth_leap()} returns a boolean vector,
\code{as.Date()} returns a Date object,
\code{as.numeric()} returns number of date since 1970-01-01 GC (1962-04-23 EC), and
\code{as.character()} returns formatted character date.
}
\description{
Small helper functions.
}
\examples{
is_eth_leap(2011)
}
